fix: apply review changes (2)
