+++ deck.gl/examples/pure-js/vue/src/components/header-component/header.component.scss
@@ -6,7 +6,7 @@
   height: 100%;
   width: 100%;
   display: flex;
-  background-color: #036fe2;
+  background-color: #036fe2;;
   padding: 0 1.5rem;
 
   .title {
