+++ deck.gl/examples/pure-js/vue/src/store/map/constants.js
@@ -2,7 +2,10 @@ export const MODULE_NAME = 'map'
 
 export const ACTIONS = {}
 
-export const GETTERS = {}
+export const GETTERS = {
+  GET_CREDENTIALS: 'getCredentials',
+  GET_BASEMAP: 'getBasemap'
+}
 
 export const MUTATIONS = {
   SET_MAP_LOADED: 'setMapLoaded',
