+++ deck.gl/examples/pure-js/vue/src/store/map/state.js
@@ -1,6 +1,7 @@
 /**
  *  State module file
 **/
+import { BASEMAP } from '@deck.gl/carto'
 
 export const initialViewState = () => ({
   longitude: -97.2,
@@ -9,6 +10,11 @@ export const initialViewState = () => ({
 })
 
 export const state = {
+  credentials: {
+    username: 'public',
+    apiKey: 'default_public'
+  },
+  basemap: BASEMAP.VOYAGER,
   mapLoaded: false,
   viewState: initialViewState(),
   viewportFeatures: []
